+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.aplicacaoprojeto"
+        applicationId "com.mindoverflow.scoutshub"
         minSdkVersion 19
         targetSdkVersion 30
         versionCode 1
