+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,67 +7,67 @@
 
     <fragment
         android:id="@+id/navigation_estatisticas"
-        android:name="com.example.aplicacaoprojeto.ui.EstatisticasFragment"
+        android:name="com.mindoverflow.scoutshub.ui.EstatisticasFragment"
         android:label="@string/title_estatisticas"
         tools:layout="@layout/fragment_estatisticas" />
     <!--
         <fragment
             android:id="@+id/navigation_atividade"
-                android:name="com.example.aplicacaoprojeto.ui.AtividadesFragment"
+                android:name="scoutshubvidadesFragment"
             android:label="@string/title_atividade"
             tools:layout="@layout/fragment_atividades" />
 
         <fragment
             android:id="@+id/navigation_material"
-            android:name="com.example.aplicacaoprojeto.ui.MateriaisFragment"
+            android:name="scoutshuberiaisFragment"
             android:label="@string/title_material"
             tools:layout="@layout/fragment_materiais" />
 
         <fragment
             android:id="@+id/navigation_materialUsadoAtividade"
-            android:name="com.example.aplicacaoprojeto.ui.MateriaisUsadosAtividadeFragment"
+            android:name="scoutshuberiaisUsadosAtividadeFragment"
             android:label="@string/title_materialUsadoAtividade"
             tools:layout="@layout/fragment_materiais_usado_atividade" />
 
         <fragment
             android:id="@+id/navigation_perfil"
-            android:name="com.example.aplicacaoprojeto.ui.PerfisFragment"
+            android:name="scoutshubfisFragment"
             android:label="@string/title_perfil"
             tools:layout="@layout/fragment_perfis" />
 
         <fragment
             android:id="@+id/navigation_equipa"
-            android:name="com.example.aplicacaoprojeto.ui.EquipasFragment"
+            android:name="scoutshubipasFragment"
             android:label="@string/title_equipa"
             tools:layout="@layout/fragment_equipas" />
 
         <fragment
             android:id="@+id/navigation_seccao"
-            android:name="com.example.aplicacaoprojeto.ui.SeccoesFragment"
+            android:name="scoutshubcoesFragment"
             android:label="@string/title_seccao"
             tools:layout="@layout/fragment_seccoes" />
 
         <fragment
             android:id="@+id/navigation_utilizador"
-            android:name="com.example.aplicacaoprojeto.ui.UtilizadoresFragment"
+            android:name="scoutshublizadoresFragment"
             android:label="@string/title_utilizador"
             tools:layout="@layout/fragment_utilizadores" />
 
         <fragment
             android:id="@+id/navigation_tipo_utilizador"
-            android:name="com.example.aplicacaoprojeto.ui.TiposUtilizadorFragment"
+            android:name="scoutshubosUtilizadorFragment"
             android:label="@string/title_tipo_utilizador"
             tools:layout="@layout/fragment_tipos_utilizador" />
 
         <fragment
             android:id="@+id/navigation_instrucao"
-            android:name="com.example.aplicacaoprojeto.ui.InstrucoesFragment"
+            android:name="scoutshubtrucoesFragment"
             android:label="@string/title_instrucoes"
             tools:layout="@layout/fragment_instrucoes" />
 
         <fragment
             android:id="@+id/navigation_instrucao_utilizador"
-            android:name="com.example.aplicacaoprojeto.ui.InstrucoesUtilizadorFragment"
+            android:name="scoutshubtrucoesUtilizadorFragment"
             android:label="@string/title_instrucao_utilizador"
             tools:layout="@layout/fragment_instrucoes_utilizador" />
     -->
